refactor: split discover tool into helper modules
