+++ src/tunacode/tools/utils/discover_terms.py
@@ -1,0 +1,213 @@
+"""Lexical vocabularies and expansion maps for discover search heuristics."""
+
+from __future__ import annotations
+
+SOURCE_EXTENSIONS = frozenset(
+    {
+        ".py",
+        ".ts",
+        ".tsx",
+        ".js",
+        ".jsx",
+        ".go",
+        ".rs",
+        ".java",
+        ".rb",
+        ".php",
+        ".c",
+        ".cpp",
+        ".h",
+        ".hpp",
+        ".cs",
+        ".swift",
+        ".kt",
+        ".scala",
+        ".lua",
+        ".sh",
+        ".bash",
+        ".zsh",
+        ".md",
+        ".txt",
+        ".yaml",
+        ".yml",
+        ".toml",
+        ".json",
+        ".ini",
+        ".cfg",
+        ".conf",
+        ".sql",
+    }
+)
+
+_PYTHON_KEYWORDS = frozenset(
+    {
+        "False",
+        "None",
+        "True",
+        "and",
+        "as",
+        "assert",
+        "async",
+        "await",
+        "break",
+        "class",
+        "continue",
+        "def",
+        "del",
+        "elif",
+        "else",
+        "except",
+        "finally",
+        "for",
+        "from",
+        "global",
+        "if",
+        "import",
+        "in",
+        "is",
+        "lambda",
+        "nonlocal",
+        "not",
+        "or",
+        "pass",
+        "raise",
+        "return",
+        "try",
+        "while",
+        "with",
+        "yield",
+    }
+)
+
+_JS_KEYWORDS = frozenset(
+    {
+        "const",
+        "let",
+        "var",
+        "function",
+        "class",
+        "export",
+        "default",
+        "import",
+        "from",
+        "return",
+        "if",
+        "else",
+        "for",
+        "while",
+        "do",
+        "switch",
+        "case",
+        "break",
+        "continue",
+        "new",
+        "this",
+        "super",
+        "async",
+        "await",
+        "try",
+        "catch",
+        "finally",
+        "throw",
+        "typeof",
+        "instanceof",
+        "void",
+        "delete",
+        "in",
+        "of",
+        "yield",
+    }
+)
+
+ALL_KEYWORDS = _PYTHON_KEYWORDS | _JS_KEYWORDS
+
+CONCEPT_EXPANSIONS: dict[str, list[str]] = {
+    "auth": [
+        "auth",
+        "login",
+        "session",
+        "token",
+        "jwt",
+        "oauth",
+        "credential",
+        "permission",
+        "role",
+        "passport",
+        "sso",
+    ],
+    "database": [
+        "db",
+        "database",
+        "model",
+        "schema",
+        "migration",
+        "orm",
+        "query",
+        "repository",
+        "entity",
+    ],
+    "api": [
+        "api",
+        "route",
+        "endpoint",
+        "controller",
+        "handler",
+        "middleware",
+        "request",
+        "response",
+    ],
+    "test": ["test", "spec", "fixture", "mock"],
+    "config": ["config", "settings", "env", "constant"],
+    "error": ["error", "exception", "fallback", "retry"],
+    "cache": ["cache", "redis", "memcache", "store"],
+    "queue": ["queue", "worker", "job", "task", "celery"],
+    "ontolog": [
+        "relation",
+        "graph",
+        "edge",
+        "node",
+        "entity",
+        "schema",
+        "hierarchy",
+        "taxonomy",
+    ],
+    "type": ["type", "schema", "interface", "protocol", "contract"],
+    "tool": ["tool", "command", "action", "executor"],
+    "agent": ["agent", "loop", "step", "run", "execute"],
+    "ui": ["ui", "widget", "panel", "render", "display", "screen"],
+    "prompt": ["prompt", "system", "template", "instruction"],
+    "compaction": ["compact", "compaction", "summary", "summarize", "truncate"],
+}
+
+NOISE_WORDS = frozenset(
+    {
+        "where",
+        "what",
+        "how",
+        "the",
+        "and",
+        "are",
+        "all",
+        "code",
+        "related",
+        "find",
+        "show",
+        "module",
+        "file",
+        "defined",
+        "does",
+        "this",
+        "that",
+        "which",
+        "with",
+        "for",
+        "from",
+        "about",
+        "look",
+        "get",
+        "used",
+        "using",
+        "implement",
+        "implementation",
+    }
+)
